refactor(graphql): split schema.py
